Foundation Built
Added the foundations to my map and the implementation of my castle Zork game. Rooms names and exits changed to cater to the story of the game. Added Courtyard to center of map which will be used as a death room as if the player enters this room he will fall to his death.